move commandmanager to src
